style(x3f-core): fix clippy 1.95 lints (-D warnings)
CI's `clippy --workspace --all-targets -- -D warnings` runs on stable,
now 1.95, which surfaced 56 lints (previously masked behind the opencv
build break). All are behaviour-preserving:

  - 48x excessive_precision: trailing-zero/`eNN` literal trims in
    tone_curves.rs. The lint rewrites each literal to the shortest decimal
    that round-trips to the *same* float bits, so DNG output is unchanged
    and the tier-2 parity MD5s hold.
  - io_other_error (new in 1.95) -> io::Error::other
  - needless_range_loop, redundant slicing, field_reassign_with_default,
    no-effect ops, write!-with-trailing-newline, redundant f64->f64 cast.

Verified locally on rustc/clippy 1.95.0: clippy clean under -D warnings,
cargo fmt --check clean, cargo test --workspace green (tier-2/3 skip with
no corpus).

+++ crates/x3f-core/src/output/dng/mod.rs
@@ -329,7 +329,7 @@ fn populate_raw_ifd(
                 .levels
                 .black
                 .iter()
-                .map(|&b| (((b as f64) * denom as f64).round().max(0.0) as u32, denom))
+                .map(|&b| ((b * denom as f64).round().max(0.0) as u32, denom))
                 .collect(),
         ),
     );
